+++ include/math/ne_math_m3_inline.h
@@ -11,14 +11,14 @@
  *                                                                       *
  *************************************************************************/
 
-NEINLINE neV3 &neM3::operator[](s32 I) {
+NEINLINE neV3 &neM3::operator[](size_t I) {
     ASSERT(I >= 0);
     ASSERT(I < 3);
 
     return M[I];
 }
 
-NEINLINE neV3 neM3::operator[](s32 I) const {
+NEINLINE neV3 neM3::operator[](size_t I) const {
     ASSERT(I >= 0);
     ASSERT(I < 3);
