+++ include/math/ne_math_v4_inline.h
@@ -79,7 +79,7 @@ NEINLINE f32 neV4::Dot(const neV4 &V) const {
 
 //=========================================================================
 
-NEINLINE f32 &neV4::operator[](s32 I) {
+NEINLINE f32 &neV4::operator[](size_t I) {
     ASSERT(I >= 0);
     ASSERT(I <= 3);
     return ((f32 *) this)[I];
